zshrc: use keychain if available and set GPG_TTY
Without setting GPG_TTY when using git commit with signing it will ask
for the passphrase on the TTY where the GPG agent was started up. This
is weird if you are committing from a different terminal window from
where it was started up. If keychain is installed then use it to manage
the GPG and SSH agents.